% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brnsims.R
\name{brnsims}
\alias{brnsims}
\title{Simulate non-stationary Brown-Resnick process}
\usage{
brnsims(reps, locs, kappa, lambda, tau, centre = NULL)
}
\arguments{
\item{reps}{Number of realisations, \eqn{n}.}

\item{locs}{A \eqn{p} by 2 matrix of coordinates.}

\item{kappa}{Value of \eqn{\kappa}. See \code{help(gamsv)}.}

\item{lambda}{Value of \eqn{\lambda}. See \code{help(gamsv)}.}

\item{tau}{A \eqn{p} by \eqn{p} matrix of correlation values.}

\item{centre}{Vector of length 2 giving coordinates of centre of non-stationarity for semivariogram. If \code{NULL}, stationary semivariogram used. See \code{help(gamsv)}.}
}
\value{
\eqn{n} by \eqn{p} matrix
}
\description{
Simulate a non-stationary (or stationary) Brown-Resnick process using the 'stopping rule'
methodology of Dieker and Mikosch (2015). Non-stationary Brown-Resnick processes are simulated
using the non-stationary semivariogram. detailed in the help file for \code{gamsv}.
}
\examples{
##Creating correlation values to simulate non-stationary Brown-Resnick process.

lambda<-2
centre<-c(0,0)
kappa<-0.8

n.grid<-8
sim.coords<-as.matrix(expand.grid(seq(-1,1,length=n.grid),seq(-1,1,length=n.grid)))

p<-dim(sim.coords)[1]
tau<-matrix(NA,nrow=p,ncol=p)

for(i in 1:p){
for(j in 1:p){
tau[i,j]<-gamsv(s1=sim.coords[i,],s2=c(0,0),lam=lambda,kap=kappa,centre=NULL)+
gamsv(s1=sim.coords[j,],s2=c(0,0),lam=lambda,kap=kappa,centre=NULL)-
gamsv(s1=sim.coords[i,],s2=sim.coords[j,],lam=lambda,kap=kappa,centre=centre)
}
}

##Simulates 10 realisations of non-stationary BR process. 

Sim<-brnsims(reps=10,locs=sim.coords,kappa=kappa,lambda=lambda,centre=centre,tau=tau)


}
\references{
Dieker and Mikosch (2015) Extremes, 18(2):301-314,
(\href{https://doi.org/10.1007/s10687-015-0214-4}{doi})
}
