% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi3emp.R
\name{chi3.emp}
\alias{chi3.emp}
\title{Empirical estimate of triple-wise chi}
\usage{
chi3.emp(U, V, W, q)
}
\arguments{
\item{U}{Vector of length \eqn{n} of standard uniform variables.}

\item{V}{Vector of length \eqn{n} of standard uniform variables.}

\item{W}{Vector of length \eqn{n} of standard uniform variables.}

\item{q}{Value of exceedance probability used in estimation.}
}
\value{
Estimate of triple-wise chi for \eqn{(U,V,W)}.
}
\description{
Calculates empirical estimate of triple-wise chi for \code{(U,V,W)}.
}
\examples{

data(Aus_Heat)
Z<-Aus_Heat$Temp.

unif<-function(x) rank(x)/(length(x)+1)
Z_U<-Z
for(i in 1:dim(Z_U)[2]) Z_U[,i]<-unif(Z[,i]) # Transform to uniform margins

q<-0.95

ind.triple<-c(1,2,3) ##Denotes indices of triple for which triple-wise chi calculated.

chi3.emp(Z_U[,ind.triple[1]],Z_U[,ind.triple[2]],Z_U[,ind.triple[3]],q)

}
