% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/High_Precip_Output.R
\docType{data}
\name{High_Precip_Output}
\alias{High_Precip_Output}
\title{Highlands Precipitation outputs}
\format{
A list with 5 elements:
\describe{
\item{likG.IMSP}{Optim output from fitting Inverted Brown-Resnick model to G-plane sampling locations. See \code{nllIMSPexp}.}
\item{likG.MSP}{Optim output from fitting Brown-Resnick model to G-plane sampling locations. See \code{nllMSPexp}.}
\item{likD.IMSP}{Optim output from fitting Inverted Brown-Resnick model to D-plane sampling locations. See \code{nllIMSPexp}.}
\item{likD.MSP}{Optim output from fitting Brown-Resnick model to D-plane sampling locations. See \code{nllMSPexp}.}
\item{sdf}{D-plane transformation spline parameters. See \code{sdf.heur.EXTR}.}
}
}
\usage{
data(High_Precip_Output)
}
\description{
Outputs from deformations and model fitting for \code{data(High_Precip)}.
}
\examples{
data(High_Precip_Output)
}
\references{
Lowe et al. (2018) Met Office, Exter, UK,
(\href{https://www.metoffice.gov.uk/pub/data/weather/uk/ukcp18/science-reports/UKCP18-Overview-report.pdf}{pdf})
}
\keyword{output}
