% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_reject.R
\name{m.reject}
\alias{m.reject}
\title{A heuristic for deciding if initial anchor points have enough spread.}
\usage{
m.reject(m.ind, Gcoords, sphere.dis = FALSE)
}
\arguments{
\item{m.ind}{A vector of indices denoting the anchor points in \code{Gcoords}.}

\item{Gcoords}{A \code{d} by 2 matrix of G-plane sampling locations.}

\item{sphere.dis}{Is Spherical distance or Euclidean distance used?}
}
\value{
\code{1} if anchor points are rejected, \code{0} otherwise.
}
\description{
This heuristic ensures that chosen anchor points are not along a straight line and,
if \code{length(m.ind)<5}, that the maximum distance between the initial anchor points in the
G-plane is sufficiently large i.e. over \eqn{60\%} of the maximum pairwise distance of all sampling locations in the G-plane.
}
\examples{
data("Aus_Heat")

Gcoords<-Aus_Heat$coords
# Set number of anchor points
m<-10
# Sample anchor points
m.ind<-sample(1:dim(Gcoords)[1],m,replace=FALSE)

reject<-m.reject(m.ind,Gcoords,sphere.dis=TRUE)
print(reject)
}
