% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindDcoords.R
\name{returnDcoord}
\alias{returnDcoord}
\alias{returnDGridcoord}
\title{Return D-plane coordinates}
\usage{
returnDcoord(par, Gcoords, whichm, sphere.dis = F)

returnDGridcoord(par, gridcoord, Gcoords, whichm, sphere.dis = F)
}
\arguments{
\item{par}{Spline parameter values calculated using either \code{FindSplineParFNEXTR} or \code{FindSplineParFNCOR} or \code{sdf.heur.Cor} or \code{sdf.heur.EXTR}.}

\item{Gcoords}{A \code{d} by 2 matrix of G-plane sampling locations.}

\item{whichm}{A vector of length \code{m < d} giving the indices of the anchor points in \code{Gcoords}. Must be the same as used to find \code{par}.}

\item{sphere.dis}{Is Spherical distance or Euclidean distance used?}

\item{gridcoord}{A \code{K} by 2 matrix of any coordinates in the G-plane.}
}
\value{
\code{returnDcoord} returns a \code{d} by 2 matrix of coordinates. \code{returnDGridcoord} returns a \code{K} by 2 matrix of coordinates.
}
\description{
Returns the D-plane coordinates given spline parameter values calculated using \code{FindSplineParFNEXTR} or \code{FindSplineParFNCOR}.
}
